Move routing code to separate js file.
